Move document rendering out of models into DocumentRenderService
Rendering a handlebars template into HTML/XML is not data access, and it
needed the platform to do it. Rather than passing a platform into the
models, the rendering moves up into the api app.

New DocumentRenderService takes buildContext, getRenderedHtml and
getRenderedHtmlForTemplate off Document, and buildContext and
getRenderedXml off DocumentTemplate. Together that is -164 lines from
@stamhoofd/models, and neither model resolves a platform any more.

DocumentTemplate.buildContext was the clearest case: it calls
buildAll({ generateNumbers: true }) and then saves documents to make the
numbering strictly increasing. That is a write driven by an export
operation, not something a template row should own.

Not folded into the existing DocumentService: that class is purely event
bus wiring (a listen() that watches Group updates). Rendering has a
different lifetime, different dependencies and different callers, so
merging would have been name matching rather than cohesion. The real
precedent is InvoicePdfService, which sits alongside and has this exact
shape.

Production no-op. The two error paths differ and are preserved verbatim:
getRenderedHtml logs and returns null, while getRenderedXml rethrows
SimpleError/SimpleErrors and only swallows other errors. Flattening that
asymmetry is the obvious way to get this wrong, so it has a test and a
mutation check -- removing the rethrow fails exactly that test.

Also preserved: the xmlExport early return, and the fact that
DocumentTemplate's renumbering does not skip locked documents the way
buildAll does. That asymmetry is what makes the rethrow reachable at all.

Tests were written against the model methods first and confirmed passing
there, then re-pointed at the service. 9 tests: the platform logo
fallback chain, null-and-log on a context failure, null on a missing
template, the renumber-and-save side effect, the SimpleError rethrow,
null for a generic error, and the no-xmlExport early return. Both error
tests drive real failures rather than mocks -- an invalid field id, and
a locked document that cannot be renumbered.

Third caller updated beyond the two endpoints: the api e2e suite in
tests/e2e/documents.test.ts called getRenderedHtml directly.
